Bug fixes
version v0.2.4

+++ lwm2m-client/m2mdevice.h
@@ -112,7 +112,7 @@ public:
      * @param value, Value to be set on the resource, in Integer format.
      * @return M2MResource if created successfully else NULL.
      */
-    M2MResource* create_resource(DeviceResource resource, uint32_t value);
+    M2MResource* create_resource(DeviceResource resource, int64_t value);
 
     /**
      * @brief Creates a new resource instance for given resource enum.
@@ -122,7 +122,7 @@ public:
      * @param value, Value to be set on the resource, in Integer format.
      * @return M2MResourceInstance if created successfully else NULL.
      */
-    M2MResourceInstance* create_resource_instance(DeviceResource resource, uint32_t value,
+    M2MResourceInstance* create_resource_instance(DeviceResource resource, int64_t value,
                                                   uint16_t instance_id);
 
     /**
@@ -178,7 +178,7 @@ public:
      * @return True if successfully set else false.
      */
     bool set_resource_value(DeviceResource resource,
-                            uint32_t value,
+                            int64_t value,
                             uint16_t instance_id = 0);
 
     /**
@@ -202,7 +202,7 @@ public:
      * @param instance_id, Instance Id of the resource, default is 0
      * @return Value associated with that key, if key is not valid it returns -1.
      */
-    uint32_t resource_value_int(DeviceResource resource,
+    int64_t resource_value_int(DeviceResource resource,
                                uint16_t instance_id = 0) const;
 
     /**
